requirements.txt
aiohttp==3.2.1
aiohttp-jinja2==1.1.1
aiopg==0.16.0
aiotask-context==0.6.0
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
psycopg2-binary==2.8.2
typing-extensions==3.7.2
uvloop==0.9.1
yarl==1.3.0
